try to fix float precision

+++ destination/schema/kafkaconnect.go
@@ -33,7 +33,7 @@ var schemaTypes = map[string]avro.Type{
 	"STRING":  avro.String,
 	"BOOLEAN": avro.Boolean,
 	"FLOAT64": avro.Double,
-	"FLOAT32": avro.Float,
+	"FLOAT32": avro.Double,
 }
 
 type (
